requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
jinja2
python-dotenv
openai